% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_indices.R
\name{a2i}
\alias{a2i}
\title{Convert an alignment position to an index}
\usage{
a2i(sti, ai, nstpist, napst)
}
\arguments{
\item{sti}{the species tree index, a value from 1 to and including 2}

\item{ai}{the alignment index,
the ai-th alignment of that species tree,
a value from 1 to and including
the number of alignments per species tree}

\item{nstpist}{the number of species trees per incipient species tree}

\item{napst}{the number of alignments per species tree}
}
\value{
the plain index
}
\description{
Convert an alignment position to an index
}
\author{
Richel Bilderbeek
}

