% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alignments.R
\name{add_alignments}
\alias{add_alignments}
\title{Add an alignment to a file}
\usage{
add_alignments(filename, verbose = TRUE)
}
\arguments{
\item{filename}{Parameter filename}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
Nothing, modifies the parameter file
}
\description{
Add an alignment to a file
}
\author{
Richel Bilderbeek
}

