% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pbd_output.R
\name{add_pbd_output}
\alias{add_pbd_output}
\title{Adds a pbd_sim result to a file}
\usage{
add_pbd_output(filename)
}
\arguments{
\item{filename}{Parameter filename}
}
\value{
Nothing, modifies the parameter file
}
\description{
Adds a pbd_sim result to a file
}
\author{
Richel Bilderbeek
}

