% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posteriors.R
\name{add_posteriors}
\alias{add_posteriors}
\title{Add BEAST2 posteriors to a file}
\usage{
add_posteriors(filename)
}
\arguments{
\item{filename}{Parameter filename}
}
\value{
Nothing, modifies the parameter file
}
\description{
Add BEAST2 posteriors to a file
}
\author{
Richel Bilderbeek
}

