% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_species_trees_with_outgroup.R
\name{add_species_trees}
\alias{add_species_trees}
\title{Add a species tree with/without outgroup to a file}
\usage{
add_species_trees(filename, verbose, add_outgroup)
}
\arguments{
\item{filename}{Parameter filename}

\item{verbose}{give verbose output, should be TRUE or FALSE}

\item{add_outgroup}{add an outgroup to the phylogeny, should be TRUE or FALSE}
}
\value{
Nothing, modifies the parameter file
}
\description{
Add a species tree with/without outgroup to a file
}
\author{
Richel Bilderbeek
}

