% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_species_trees_with_outgroup.R
\name{add_species_trees_with_outgroup}
\alias{add_species_trees_with_outgroup}
\title{Add a species tree with outgroup to a file}
\usage{
add_species_trees_with_outgroup(filename)
}
\arguments{
\item{filename}{Parameter filename}
}
\value{
Nothing, modifies the parameter file
}
\description{
Add a species tree with outgroup to a file
}
\author{
Richel Bilderbeek
}

