% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_to_beast_input_file.R
\name{alignment_to_beast_input_file}
\alias{alignment_to_beast_input_file}
\title{Convert an alignment and parameters to a BEAST XML input file}
\usage{
alignment_to_beast_input_file(alignment_dnabin, nspp, rng_seed = 42,
  beast_filename, temp_fasta_filename, verbose = FALSE)
}
\arguments{
\item{alignment_dnabin}{the DNA alignment of type DNAbin}

\item{nspp}{the number of states in a BEAST2 MCMC chain,
typically there is one state per one thousand generations}

\item{rng_seed}{random number generator seed}

\item{beast_filename}{the filename of the XML BEAST2 input file created}

\item{temp_fasta_filename}{the name of a temporary file created}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
Nothing, creates a file called 'beast_filename'
}
\description{
Convert an alignment and parameters to a BEAST XML input file
}
\examples{

  # Simulate a phylogeny
  phylogeny <- ape::rcoal(n = 5)

  # Simulate an alignment from that phylogeny
  alignment <- convert_phylogeny_to_alignment(
    phylogeny = phylogeny,
    sequence_length = 10
  )

  # Create a BEAST2 input file
  beast_xml_input_file <- "alignment_to_beast_input_file_example.xml"
  fasta_filename <- "alignment_to_beast_input_file_examp.fasta"
  alignment_to_beast_input_file(
    alignment_dnabin = alignment,
    nspp = 10,
    rng_seed = 42,
    beast_filename = beast_xml_input_file,
    temp_fasta_filename = fasta_filename
  )

  # Check that the BEAST2 input file really exists
  testit::assert(file.exists(beast_xml_input_file))

}
\author{
Richel Bilderbeek
}

