% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_to_beast_posterior.R
\name{alignment_to_beast_posterior}
\alias{alignment_to_beast_posterior}
\title{Convert an alignment to a BEAST2 XML input file}
\usage{
alignment_to_beast_posterior(alignment_dnabin, mcmc_chainlength, base_filename,
  rng_seed = 42, beast_jar_path = find_beast_jar_path(),
  skip_if_output_present = FALSE, verbose = FALSE)
}
\arguments{
\item{alignment_dnabin}{The alignmnet}

\item{mcmc_chainlength}{The length of the MCMC chain BEAST2 will generate}

\item{base_filename}{The base of the filename (the part without the extension)}

\item{rng_seed}{The random number generator seed used by BEAST2}

\item{beast_jar_path}{Where the jar 'beast.jar' can be found}

\item{skip_if_output_present}{skip if output files are present, else remove these and start a new BEAST2 run}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
the phylogenies of the BEAST2 posterior
}
\description{
Convert an alignment to a BEAST2 XML input file
}
\author{
Richel Bilderbeek
}

