% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments_same_distr.R
\name{alignments_same_distr_nltt_stat}
\alias{alignments_same_distr_nltt_stat}
\title{Tests if the nLTT statistics between alignments
are of the same distribution}
\usage{
alignments_same_distr_nltt_stat(df)
}
\arguments{
\item{df}{data frame with column names 'filename',
 'sti' (Species Tree Index), 'ai' (Alignment Index), 'pi' (Posterior
Index), 'si' (State Index) and 'nltt_stat' (the nLTT statistic between
that posterior state its phylogeny and the sampled species tree)}
}
\value{
data frame
}
\description{
Tests if the nLTT statistics between alignments
are of the same distribution
}
\examples{
  nltt_stats <- read_collected_nltt_stats()
  df <- alignments_same_distr_nltt_stat(nltt_stats)
  testit::assert(all(names(df) == c("filename", "sti", "same_distr")))
}

