% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_multi.R
\name{analyse_multi}
\alias{analyse_multi}
\title{Analyse multiple files}
\usage{
analyse_multi(filenames)
}
\arguments{
\item{filenames}{Parameter filenames}
}
\value{
Nothing?
}
\description{
Analyse multiple files
}
\author{
Richel Bilderbeek
}

