% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_single.R
\name{analyse_single}
\alias{analyse_single}
\title{Analyse a single file}
\usage{
analyse_single(filename)
}
\arguments{
\item{filename}{Parameter filename}
}
\value{
Nothing?
}
\description{
Analyse a single file
}
\author{
Richel Bilderbeek
}

