% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_beast_replicates_from_same_distribution.R
\name{are_beast_replicates_from_same_distribution}
\alias{are_beast_replicates_from_same_distribution}
\title{Tests if the nLTT statistics are of the same distribution}
\usage{
are_beast_replicates_from_same_distribution(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame,
}
\description{
Tests if the nLTT statistics are of the same distribution
}

