% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_from_same_distribution.R
\name{are_from_same_distribution}
\alias{are_from_same_distribution}
\title{Checks if two samples are from the same distribution
with a Mann-Whitney U test}
\usage{
are_from_same_distribution(a, b, alpha = 0.05)
}
\arguments{
\item{a}{samples}

\item{b}{samples}

\item{alpha}{significance level, is 0.05 by default}
}
\value{
TRUE or FALSE, or NA if cannot be calculated
}
\description{
Checks if two samples are from the same distribution
with a Mann-Whitney U test
}

