% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_posteriors.R
\name{are_identical_posteriors}
\alias{are_identical_posteriors}
\title{Determines if the input is a Cer2016 posterior,}
\usage{
are_identical_posteriors(p, q, verbose = FALSE)
}
\arguments{
\item{p}{the first Cer2016 posterior}

\item{q}{the second Cer2016 posterior}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
TRUE or FALSE
}
\description{
Determines if the input is a Cer2016 posterior,
}
\author{
Richel Bilderbeek
}

