% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_trees_posteriors.R
\name{are_identical_trees_posteriors}
\alias{are_identical_trees_posteriors}
\title{Determines if the input is a BEAST2 posterior,
as parsed by olli's rBEAST package its function beast2out.read.trees}
\usage{
are_identical_trees_posteriors(p, q, verbose = FALSE)
}
\arguments{
\item{p}{the first posterior}

\item{q}{the second posterior}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
TRUE or FALSE
}
\description{
Determines if the input is a BEAST2 posterior,
as parsed by olli's rBEAST package its function beast2out.read.trees
}
\author{
Richel Bilderbeek
}

