% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_sorter.R
\name{branch_sorter}
\alias{branch_sorter}
\title{sorts branches for the NRBS analysis}
\usage{
branch_sorter(filename)
}
\arguments{
\item{filename}{name of the file with the trees that are to be sorted?}
}
\value{
a sorted tree1 and tree2
}
\description{
sorts branches for the NRBS analysis
}
\author{
Femke Thon
}

