% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_sorter.R
\name{branch_sorter_from_file}
\alias{branch_sorter_from_file}
\title{sorts branches for the NRBS analysis}
\usage{
branch_sorter_from_file(filename)
}
\arguments{
\item{filename}{name of a file}
}
\value{
a sorted tree1 and tree2
}
\description{
sorts branches for the NRBS analysis
}
\author{
Femke Thon
}

