% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ess.R
\name{calc_act}
\alias{calc_act}
\title{Calculate the autocorrelation time}
\usage{
calc_act(trace, sampleInterval)
}
\arguments{
\item{trace}{the values}

\item{sampleInterval}{the interval in timesteps between samples}
}
\value{
the autocorrelation time
}
\description{
Calculate the autocorrelation time
}
\author{
Richel Bilderbeek
}

