% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nltt_stats.R
\name{calc_nltt_stats}
\alias{calc_nltt_stats}
\title{Calculate the nLTT statistic between a focal phylogeny
and multiple others}
\usage{
calc_nltt_stats(phylogeny, others, dt)
}
\arguments{
\item{phylogeny}{the focal phylogeny}

\item{others}{the other phylogenies}

\item{dt}{resolution of the analysis, must be between zero and one}
}
\value{
A data frame
}
\description{
Calculate the nLTT statistic between a focal phylogeny
and multiple others
}
\author{
Richel Bilderbeek
}

