% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nltt_stats_from_file.R
\name{calc_nltt_stats_from_file}
\alias{calc_nltt_stats_from_file}
\title{Calculates the nLTT statistics from a file}
\usage{
calc_nltt_stats_from_file(filename)
}
\arguments{
\item{filename}{the name of a file}
}
\value{
a distribution of nLTT statistics
}
\description{
Calculates the nLTT statistics from a file
}
\examples{
  nltt_stats <- calc_nltt_stats_from_file(
    filename = find_path("toy_example_1.RDa")
  )
  expected_names <- c("sti", "ai", "pi", "si", "nltt_stat")
  testit::assert(names(nltt_stats) == expected_names)
}
\author{
Richel Bilderbeek
}

