% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nltt_stats_from_files.R
\name{calc_nltt_stats_from_files}
\alias{calc_nltt_stats_from_files}
\title{Calculates the nLTT statistics from multiple files}
\usage{
calc_nltt_stats_from_files(filenames)
}
\arguments{
\item{filenames}{the names of the files}
}
\value{
a distribution of nLTT statistics
}
\description{
Calculates the nLTT statistics from multiple files
}
\examples{
  nltt_stats <- calc_nltt_stats_from_files(
    filenames = c(
      find_path("toy_example_1.RDa"),
      find_path("toy_example_2.RDa")
    )
  )
  expected_names <- c("filename", "sti", "ai", "pi", "si", "nltt_stat")
  testit::assert(names(nltt_stats) == expected_names)
}
\author{
Richel Bilderbeek
}

