% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_nltt_stats.R
\name{cast_nltt_stats}
\alias{cast_nltt_stats}
\title{Of a dataframe, cast the multiple nltt_stat values per state index
to one row}
\usage{
cast_nltt_stats(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a data frame
}
\description{
Of a dataframe, cast the multiple nltt_stat values per state index
to one row
}

