% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_progress.R
\name{check_progress}
\alias{check_progress}
\title{Function that shows how much parameter files are present,
and how far the pipeline has gotten}
\usage{
check_progress(folder = "~/Slurm/")
}
\arguments{
\item{folder}{The folder containing the parameter files}
}
\value{
A data table showing the progress
}
\description{
Function that shows how much parameter files are present,
and how far the pipeline has gotten
}
\author{
Richel Bilderbeek
}

