% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_esses.R
\name{collect_file_esses}
\alias{collect_file_esses}
\title{Collects the ESSes of all phylogenies belonging to a
parameter file in the melted/uncast/long form}
\usage{
collect_file_esses(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A dataframe of ESSes for each posterior
}
\description{
Collects the ESSes of all phylogenies belonging to a
parameter file in the melted/uncast/long form
}
\examples{
  filename <- find_path("toy_example_3.RDa")
  df <- collect_file_esses(filename, dt = dt)
  testit::assert(nrow(df) == 8)
}

