% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_gammas.R
\name{collect_file_gammas}
\alias{collect_file_gammas}
\title{Collects the gamm statistics of all phylogenies belonging to a
parameter file in the melted/uncast/long form}
\usage{
collect_file_gammas(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of the parameter file}
}
\value{
A dataframe of gamma statistics of each phylogeny in time
}
\description{
Collects the gamm statistics of all phylogenies belonging to a
parameter file in the melted/uncast/long form
}

