% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_nltts.R
\name{collect_file_nltts}
\alias{collect_file_nltts}
\title{Collects the nLTTs of all phylogenies belonging to a
parameter file in the melted/uncast/long form}
\usage{
collect_file_nltts(filename, dt = 0.001, verbose = FALSE)
}
\arguments{
\item{filename}{name of the parameter file}

\item{dt}{the resolution of the nLTT plot,
must be in range <0,1>, default is 0.001}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A dataframe of gamma statistics of each phylogeny in time
}
\description{
Collects the nLTTs of all phylogenies belonging to a
parameter file in the melted/uncast/long form
}
\examples{
  dt <- 0.1
  filename <- find_path("toy_example_3.RDa")
  df <- collect_file_nltts(filename, dt = dt)
  testit::assert(names(df) == c("species_tree_nltts", "posterior_nltts"))
  testit::assert(names(df$species_tree_nltts)
    == c("species_tree", "t", "nltt")
  )
  testit::assert(nrow(df$species_tree_nltts) > 2)
  testit::assert(nrow(df$posterior_nltts) > 80)
}

