% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_file_nrbss.R
\name{collect_file_nrbss}
\alias{collect_file_nrbss}
\title{Collects the NRBS values between species trees and
posteriors in a parameter file in the melted/uncast/long form}
\usage{
collect_file_nrbss(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A dataframe of NRBS values between each species tree and their posteriors
}
\description{
Collects the NRBS values between species trees and
posteriors in a parameter file in the melted/uncast/long form
}
\examples{
  filename <- find_path("toy_example_3.RDa")
  df <- collect_file_nrbss(filename)
  testit::assert(nrow(df) > 2)
}

