% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_esses.R
\name{collect_files_esses}
\alias{collect_files_esses}
\title{Collects the ESSes of all phylogenies belonging to the
parameter files in the melted/uncast/long form}
\usage{
collect_files_esses(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter files}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A dataframe of ESSes for all files their posterior
}
\description{
Collects the ESSes of all phylogenies belonging to the
parameter files in the melted/uncast/long form
}
\examples{
  filenames <- find_paths(c("toy_example_3.RDa", "toy_example_4.RDa"))
  df <- collect_files_esses(filenames)
  testit::assert(nrow(df) == 16)
}

