% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_n_alignments.R
\name{collect_files_n_alignments}
\alias{collect_files_n_alignments}
\title{Collects the number of alignments of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_n_alignments(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A dataframe with all number of sampled alignments of all files
}
\description{
Collects the number of alignments of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
\examples{
  filenames <- c(
    find_path("toy_example_1.RDa"),
    find_path("toy_example_3.RDa")
  )
  df <- collect_files_n_alignments(filenames)
  testit::assert(names(df) == c("filenames", "n_alignments"))
  testit::assert(nrow(df) == length(filenames))
  testit::assert(df$n_alignments == c(2, 4))
}

