% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_n_species_trees.R
\name{collect_files_n_species_trees}
\alias{collect_files_n_species_trees}
\title{Collects the number of sampled species trees of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_n_species_trees(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A dataframe with all number of sampled species trees of all files
}
\description{
Collects the number of sampled species trees of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
\examples{
  filenames <- c(
   find_path("toy_example_1.RDa"),
   find_path("toy_example_2.RDa")
 )
 df <- collect_files_n_species_trees(filenames, verbose = FALSE)
 testit::assert(names(df) == c("filenames", "n_species_trees"))
 testit::assert(nrow(df) == length(filenames))
}

