% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_n_taxa.R
\name{collect_files_n_taxa}
\alias{collect_files_n_taxa}
\title{Collects the number of taxa of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_n_taxa(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A dataframe with all number of taxa of all files
}
\description{
Collects the number of taxa of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}

