% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_nltts.R
\name{collect_files_nltts}
\alias{collect_files_nltts}
\title{Collects the nLTT statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_nltts(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A list with two dataframes of nLTTs
}
\description{
Collects the nLTT statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}

