% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_gamma_statistics.R
\name{collect_gamma_statistics}
\alias{collect_gamma_statistics}
\title{Collects the gamma statistics of all phylogenies in the
melted/uncast/long form}
\usage{
collect_gamma_statistics(phylogenies)
}
\arguments{
\item{phylogenies}{the phylogenies, supplied as either a list or a multiPhylo object, where the phylogenies are of type 'phylo'}
}
\value{
A dataframe of gamma statistics of each phylogeny in time
}
\description{
Collects the gamma statistics of all phylogenies in the
melted/uncast/long form
}

