% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_n_posteriors.R
\name{collect_n_posteriors}
\alias{collect_n_posteriors}
\title{Collect the number of posteriors in a parameter file}
\usage{
collect_n_posteriors(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
a data frame
}
\description{
Collect the number of posteriors in a parameter file
}
\examples{
 filename <- find_path("toy_example_3.RDa")
 df <- collect_n_posteriors(filename)
 testit::assert(names(df) == c("n_posteriors"))
 testit::assert(ncol(df) == 1)
 testit::assert(nrow(df) == 1)
 testit::assert(df$n_posteriors[1] == 8)
}
\author{
Richel Bilderbeek
}

