% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_n_taxa.R
\name{collect_n_taxa}
\alias{collect_n_taxa}
\title{Collects the number of taxa of all phylogenies in the
melted/uncast/long form}
\usage{
collect_n_taxa(phylogenies)
}
\arguments{
\item{phylogenies}{the phylogenies, supplied as either a list or a multiPhylo object, where the phylogenies are of type 'phylo'}
}
\value{
A dataframe of the number of taxa of each phylogeny in time
}
\description{
Collects the number of taxa of all phylogenies in the
melted/uncast/long form
}
\examples{
  phylogenies <- c(ape::rcoal(10), ape::rcoal(20))
  df <- collect_n_taxa(phylogenies)
  testit::assert(names(df) == c("n_taxa"))
  testit::assert(df == data.frame(n_taxa = c(10, 20)))
}
\author{
Richel Bilderbeek
}

