% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_parameters.R
\name{collect_parameters}
\alias{collect_parameters}
\title{Collects the parameters of its input files}
\usage{
collect_parameters(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter files}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
a data.frame, one per parameter file. If all filenames are invalid,
  a simpler data.frame is returned
}
\description{
Collects the parameters of its input files
}
\examples{
 filenames <- c(
   find_path("toy_example_1.RDa"),
   find_path("toy_example_2.RDa"),
   find_path("toy_example_3.RDa")
 )
 df <- collect_parameters(filenames)
 testit::assert(nrow(df) == 3)
}
\author{
Richel Bilderbeek
}

