% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_posterior_nllts.R
\name{collect_posterior_nltts}
\alias{collect_posterior_nltts}
\title{Collect the nLTT values of the BEAST2 posteriors}
\usage{
collect_posterior_nltts(filename, dt)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}

\item{dt}{the resolution of the nLTT plot,
must be in range <0,1>, default is 0.001}
}
\value{
a data frame
}
\description{
Collect the nLTT values of the BEAST2 posteriors
}
\author{
Richel Bilderbeek
}

