% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_species_tree_n_taxa.R
\name{collect_species_tree_n_taxa}
\alias{collect_species_tree_n_taxa}
\title{Collect the number of taxa of the species trees}
\usage{
collect_species_tree_n_taxa(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
a data frame
}
\description{
Collect the number of taxa of the species trees
}
\examples{
 filename <- find_path("toy_example_1.RDa")
 df <- collect_species_tree_n_taxa(filename)
 testit::assert(names(df) == c("n_taxa"))
 testit::assert(ncol(df) == 1)
 testit::assert(nrow(df) == 1)
}
\author{
Richel Bilderbeek
}

