% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_species_tree_nltts.R
\name{collect_species_tree_nltts}
\alias{collect_species_tree_nltts}
\title{Collect the nLTT values of the species trees with outgroup}
\usage{
collect_species_tree_nltts(filename, dt)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}

\item{dt}{the resolution of the nLTT plot,
must be in range <0,1>, default is 0.001}
}
\value{
a data frame
}
\description{
Collect the nLTT values of the species trees with outgroup
}
\examples{
  dt <- 0.1
  filename <- find_path("toy_example_3.RDa")
  df <- collect_species_tree_nltts(filename, dt = dt)
  testit::assert(names(df) == c("species_tree", "t", "nltt"))
  testit::assert(nrow(df) == 2 * (1 + (1/dt)))
}
\author{
Richel Bilderbeek
}

