% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_species_tree_with_outgroup_gamma_statistics.R
\name{collect_species_tree_with_outgroup_gamma_statistics}
\alias{collect_species_tree_with_outgroup_gamma_statistics}
\title{Collect the gamma statistics of the species trees with outgroup}
\usage{
collect_species_tree_with_outgroup_gamma_statistics(filename, dt = 0.001)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}
}
\value{
a data frame
}
\description{
Collect the gamma statistics of the species trees with outgroup
}
\author{
Richel Bilderbeek
}

