% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_species_tree_with_outgroup_nltt_values.R
\name{collect_species_tree_with_outgroup_nltt_values}
\alias{collect_species_tree_with_outgroup_nltt_values}
\title{Collect the nLTT values of the species trees with outgroup}
\usage{
collect_species_tree_with_outgroup_nltt_values(filename, dt = 0.001)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}

\item{dt}{the resolution of the nLTT plot,
must be in range <0,1>, default is 0.001}
}
\value{
a data frame
}
\description{
Collect the nLTT values of the species trees with outgroup
}
\author{
Richel Bilderbeek
}

