% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_alignment_to_beast_input_file.R
\name{convert_alignment_to_beast_input_file}
\alias{convert_alignment_to_beast_input_file}
\title{Convert an alignment and parameters to a BEAST XML input file}
\usage{
convert_alignment_to_beast_input_file(alignment_dnabin, mcmc_chainlength,
  rng_seed = 42, beast_filename, temp_fasta_filename, verbose = TRUE)
}
\arguments{
\item{alignment_dnabin}{the DNA alignment of type DNAbin}

\item{mcmc_chainlength}{the BEAST2 MCMC chain length}

\item{rng_seed}{random number generator seed}

\item{beast_filename}{the filename of the XML BEAST2 input file created}

\item{temp_fasta_filename}{the name of a temporary file created}
}
\value{
Nothing, creates a file called 'beast_filename'
}
\description{
Convert an alignment and parameters to a BEAST XML input file
}
\author{
Richel Bilderbeek
}

