% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_alignment_to_beast_posterior.R
\name{convert_alignment_to_beast_posterior}
\alias{convert_alignment_to_beast_posterior}
\title{Convert an alignment to a BEAST2 XML input file}
\usage{
convert_alignment_to_beast_posterior(alignment_dnabin, mcmc_chainlength,
  base_filename, rng_seed = 42, beast_bin_path = "",
  beast_jar_path = find_beast_jar_path(), skip_if_output_present = FALSE)
}
\arguments{
\item{alignment_dnabin}{The alignmnet}

\item{mcmc_chainlength}{The length of the MCMC chain BEAST2 will generate}

\item{base_filename}{The base of the filename (the part without the extension)}

\item{rng_seed}{The random number generator seed used by BEAST2}

\item{beast_bin_path}{Where the binary 'beast' can be found}

\item{beast_jar_path}{Where the jar 'beast.jar' can be found}

\item{skip_if_output_present}{skip if output files are present, else remove these and start a new BEAST2 run}
}
\value{
A BEAST2 posterior
}
\description{
Convert an alignment to a BEAST2 XML input file
}
\author{
Richel Bilderbeek
}

