% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_alignment_to_fasta.R
\name{convert_alignment_to_fasta}
\alias{convert_alignment_to_fasta}
\title{Convert an alignment (of type DNAbin) to a FASTA file}
\usage{
convert_alignment_to_fasta(alignment_dnabin, filename)
}
\arguments{
\item{alignment_dnabin}{DNA alignment of type DNAbin}

\item{filename}{FASTA filename the alignment will be saved to}
}
\value{
Nothing, it will create a FASTA file
}
\description{
Convert an alignment (of type DNAbin) to a FASTA file
}
\author{
Richel Bilderbeek
}

