% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_phylogeny_to_alignment.R
\name{convert_phylogeny_to_alignment}
\alias{convert_phylogeny_to_alignment}
\title{Converts a phylogeny to a random DNA alignment}
\usage{
convert_phylogeny_to_alignment(phylogeny, sequence_length, mutation_rate = 1)
}
\arguments{
\item{phylogeny}{a phylogeny}

\item{sequence_length}{the number of nucleotides to alignment
will have per taxon}

\item{mutation_rate}{the rate per nucleotide to change,
per million years}
}
\value{
an alignment
}
\description{
Converts a phylogeny to a random DNA alignment
}
\examples{
alignment <- convert_phylogeny_to_alignment(
   phylogeny = ape::rcoal(5),
   sequence_length = 10,
   mutation_rate = 1
 )
 testit::assert(is_alignment(alignment))
}
\author{
Richel Bilderbeek
}

