% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_citations_bibtex.R
\name{create_citations_bibtex}
\alias{create_citations_bibtex}
\title{Creates the full list of citations}
\usage{
create_citations_bibtex()
}
\value{
an characterVector of all citations
}
\description{
Creates the full list of citations
}
\author{
Richel Bilderbeek
}

