% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_files_article.R
\name{create_parameter_files_article}
\alias{create_parameter_files_article}
\title{Creates the parameter files in the article}
\usage{
create_parameter_files_article()
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files in the article
}
\author{
Richel Bilderbeek
}

