% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_files_timings.R
\name{create_parameter_files_timings}
\alias{create_parameter_files_timings}
\title{Creates the parameter files in the article, but
slimmed down for time measurement}
\usage{
create_parameter_files_timings()
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files in the article, but
slimmed down for time measurement
}
\author{
Richel Bilderbeek
}

