% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_files_toy_examples.R
\name{create_parameter_files_toy_examples}
\alias{create_parameter_files_toy_examples}
\title{Creates the parameter files for the toy examples}
\usage{
create_parameter_files_toy_examples()
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files for the toy examples
}
\author{
Richel Bilderbeek
}

