% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_simulation.R
\name{do_simulation}
\alias{do_simulation}
\title{Run the full simulation pipeline on a parameter file}
\usage{
do_simulation(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
Nothing. It does modify the input filename
}
\description{
Run the full simulation pipeline on a parameter file
}
\author{
Richel Bilderbeek
}

