% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_test_simulations.R
\name{do_test_simulations}
\alias{do_test_simulations}
\title{Run the full simulation pipeline on the test parameters}
\usage{
do_test_simulations(verbose = FALSE)
}
\arguments{
\item{verbose}{give verbose output, should be TRUE or FALSE}

\item{filename}{name of the parameter file}

\item{cache_beast_output}{if the BEAST2 output files exist, do not recalculate}
}
\value{
Nothing. It does modify the input filename
}
\description{
Run the full simulation pipeline on the test parameters
}
\author{
Richel Bilderbeek
}

