% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_erg}
\alias{extract_erg}
\title{Extract the ERG ('Extinction Rate of a Good species')
  parameter value from a file}
\usage{
extract_erg(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the value of the ERG parameter
}
\description{
Extract the ERG ('Extinction Rate of a Good species')
  parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  erg <- extract_erg(file)
  testit::assert(erg >= 0.0)
}
\author{
Richel Bilderbeek
}

