% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_eri}
\alias{extract_eri}
\title{Extract the ERI ('Extinction Rate of an Incipient species')
  parameter value from a file}
\usage{
extract_eri(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the value of the ERI parameter
}
\description{
Extract the ERI ('Extinction Rate of an Incipient species')
  parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  eri <- extract_eri(file)
  testit::assert(eri >= 0.0)
}
\author{
Richel Bilderbeek
}

