% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_napst}
\alias{extract_napst}
\title{Extract the number of alignments per species tree
  parameter value from a file}
\usage{
extract_napst(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the number of alignments per species tree
}
\description{
Extract the number of alignments per species tree
  parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  napst <- extract_napst(file)
  testit::assert(napst == 1)
}
\author{
Richel Bilderbeek
}

