% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_nppa}
\alias{extract_nppa}
\title{Extract the number of posteriors per alignment
  parameter value from a file}
\usage{
extract_nppa(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the number of posteriors per alignment
}
\description{
Extract the number of posteriors per alignment
  parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  nppa <- extract_nppa(file)
  testit::assert(nppa == 1)
}
\author{
Richel Bilderbeek
}

