% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_nspp}
\alias{extract_nspp}
\title{Extract the number of states per posterior
  parameter value from a file}
\usage{
extract_nspp(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the number of posteriors per alignment
}
\description{
Extract the number of states per posterior
  parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  nspp <- extract_nspp(file)
  testit::assert(nspp == 10)
}
\author{
Richel Bilderbeek
}

