% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_posteriors}
\alias{extract_posteriors}
\title{Extract the BEAST2 posterior phyogenies
  from a file}
\usage{
extract_posteriors(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the value of the ERG parameter
}
\description{
Extract the BEAST2 posterior phyogenies
  from a file
}
\author{
Richel Bilderbeek
}

