% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_scr}
\alias{extract_scr}
\title{Extract the SCR ('Speciation Completion Rate')
  parameter value from a file}
\usage{
extract_scr(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the value of the SCR parameter
}
\description{
Extract the SCR ('Speciation Completion Rate')
  parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  scr <- extract_scr(file)
  testit::assert(scr >= 0.0)
}
\author{
Richel Bilderbeek
}

