% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_sirg}
\alias{extract_sirg}
\title{Extract the SIRG ('Speciation Initial Rate of a Good species')
  parameter value from a file}
\usage{
extract_sirg(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the value of the SIRG parameter
}
\description{
Extract the SIRG ('Speciation Initial Rate of a Good species')
  parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  sirg <- extract_sirg(file)
  testit::assert(sirg >= 0.0)
}
\author{
Richel Bilderbeek
}

