% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_siri}
\alias{extract_siri}
\title{Extract the SIRI ('Speciation Initial Rate of an Incipient species')
  parameter value from a file}
\usage{
extract_siri(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the value of the SIRI parameter
}
\description{
Extract the SIRI ('Speciation Initial Rate of an Incipient species')
  parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  siri <- extract_siri(file)
  testit::assert(siri >= 0.0)
}
\author{
Richel Bilderbeek
}

