% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_has_valid_parameters.R
\name{file_has_valid_parameters}
\alias{file_has_valid_parameters}
\title{Checks if a file is a valid parameter file}
\usage{
file_has_valid_parameters(file, verbose = FALSE)
}
\arguments{
\item{file}{a file that has been openened by read_file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
TRUE or FALSE
}
\description{
Checks if a file is a valid parameter file
}
\examples{
  testit::assert(
    file_has_valid_parameters(
      file = read_file(find_path("toy_example_1.RDa"))
    )
  )
}
\author{
Richel Bilderbeek
}

