% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_paths.R
\name{find_beast_posterior_test_filename}
\alias{find_beast_posterior_test_filename}
\title{Find the path of known-to-be-valid BEAST2 posterior file}
\usage{
find_beast_posterior_test_filename()
}
\value{
the path of a known-to-be-valid BEAST2 posterior file
}
\description{
Find the path of known-to-be-valid BEAST2 posterior file
}
\examples{
  library(testit)
  my_file <- find_beast_posterior_test_filename()
  assert(file.exists(my_file))
}

