% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment.R
\name{get_alignment}
\alias{get_alignment}
\title{Extract an alignment from an index from a file}
\usage{
get_alignment(file, sti, ai)
}
\arguments{
\item{file}{A loaded parameter file}

\item{sti}{the species tree index, a value from 1 to and including 2}

\item{ai}{the alignment index,
the ai-th alignment of that species tree,
a value from 1 to and including
the number of alignments per species tree}
}
\value{
the alignment
}
\description{
Extract an alignment from an index from a file
}
\author{
Richel Bilderbeek
}

