% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment.R
\name{get_alignment_by_index}
\alias{get_alignment_by_index}
\title{Extract an alignment from an index from a file}
\usage{
get_alignment_by_index(file, i)
}
\arguments{
\item{file}{A loaded parameter file}

\item{i}{the index of the alignment}
}
\value{
the alignment
}
\description{
Extract an alignment from an index from a file
}
\author{
Richel Bilderbeek
}

