% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignments.R
\name{get_alignments}
\alias{get_alignments}
\title{Extract the alignments
  from a file}
\usage{
get_alignments(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
all alignments
}
\description{
Extract the alignments
  from a file
}
\author{
Richel Bilderbeek
}

