% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylogeny_crown_age.R
\name{get_phylogeny_crown_age}
\alias{get_phylogeny_crown_age}
\title{Obtain the crown age of a phylony}
\usage{
get_phylogeny_crown_age(phylogeny)
}
\arguments{
\item{phylogeny}{The phylogeny to obtain the crown age of}
}
\value{
the age of the phylogeny
}
\description{
Obtain the crown age of a phylony
}
\author{
Richel Bilderbeek
}

