% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{get_posterior_by_index}
\alias{get_posterior_by_index}
\title{Extract a BEAST2 posterior phyogenies
  from a file}
\usage{
get_posterior_by_index(file, posterior_index)
}
\arguments{
\item{file}{A loaded parameter file}

\item{posterior_index}{the index of the posterior}
}
\value{
the posterior
}
\description{
Extract a BEAST2 posterior phyogenies
  from a file
}
\author{
Richel Bilderbeek
}

