% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posteriors.R
\name{get_posteriors}
\alias{get_posteriors}
\title{Extract the BEAST2 posterior phyogenies
  from a file}
\usage{
get_posteriors(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
all posteriors
}
\description{
Extract the BEAST2 posterior phyogenies
  from a file
}
\author{
Richel Bilderbeek
}

