% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_tree.R
\name{get_species_tree_by_index}
\alias{get_species_tree_by_index}
\title{Extract an species_tree from an index from a file}
\usage{
get_species_tree_by_index(file, sti)
}
\arguments{
\item{file}{A loaded parameter file}

\item{sti}{the index of the species_tree}
}
\value{
the species_tree
}
\description{
Extract an species_tree from an index from a file
}
\author{
Richel Bilderbeek
}

