% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_alignments.R
\name{has_alignments}
\alias{has_alignments}
\title{Checks if alignments are present}
\usage{
has_alignments(file)
}
\arguments{
\item{file}{a file, as returned by 'read_file(filename)'}
}
\value{
a vector of TRUEs and FALSEs
}
\description{
Checks if alignments are present
}
\author{
Richel Bilderbeek
}

