% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast_posterior.R
\name{is_beast_posterior}
\alias{is_beast_posterior}
\title{Determines if the input is a BEAST2 posterior}
\usage{
is_beast_posterior(x)
}
\arguments{
\item{x}{the input}
}
\value{
TRUE or FALSE
}
\description{
Determines if the input is a BEAST2 posterior
}
\author{
Richel Bilderbeek
}

