% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distributed_normally.R
\name{is_distributed_normally}
\alias{is_distributed_normally}
\title{Sees if the values of a distribution are normally distributed
 using a Shapiro-Wilk normality test}
\usage{
is_distributed_normally(values, p_value = 0.05)
}
\arguments{
\item{values}{the values to check}

\item{p_value}{the p value}
}
\value{
TRUE when it cannot be rejected that the distribution is normal,
   FALSE when it can be rejected that the distribution is normal,
   NA when all values have been identical. The function will call
   stop if not all values are numeric or the number of values
   is not between three and five thousand
}
\description{
Sees if the values of a distribution are normally distributed
 using a Shapiro-Wilk normality test
}
\examples{

   # Set a random number generator seed for reproducible results
   set.seed(42)

   # Create a normal distribution
   nd <- rnorm(n = 1000, mean = 0.0, sd = 1.0)
   # The normal distribution should be normal
   testit::assert(is_distributed_normally(nd))

   # Create a non-normal distribution
   nnd <- runif(n = 1000, min = 0.0, max = 1.0)
   # The non-normal distribution should not be normal
   testit::assert(!is_distributed_normally(nnd))

   # Create a disribution of one value only
   r <- rep(x = 42, times = 100)
   # This one-value distribution should return NA
   testit::assert(is.na(is_distributed_normally(r)))

}
\author{
Richel Bilderbeek
}

