% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_posterior.R
\name{is_posterior}
\alias{is_posterior}
\title{Determines if the input is a BEAST2 posterior}
\usage{
is_posterior(x, verbose = FALSE)
}
\arguments{
\item{x}{the input}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
TRUE or FALSE
}
\description{
Determines if the input is a BEAST2 posterior
}
\examples{

  file <- read_file(find_path("toy_example_1.RDa"))
  posterior <- get_posterior(file = file, sti = 1, ai = 1, pi = 1)
  testit::assert(is_posterior(posterior))

}
\author{
Richel Bilderbeek
}

