% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_file.R
\name{is_valid_file}
\alias{is_valid_file}
\title{Checks if a file is a valid parameter file}
\usage{
is_valid_file(filename)
}
\arguments{
\item{filename}{the name of the file to be checked}
}
\value{
TRUE or FALSE
}
\description{
Checks if a file is a valid parameter file
}
\author{
Richel Bilderbeek
}

