% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parameters_from_file.R
\name{load_parameters_from_file}
\alias{load_parameters_from_file}
\title{Load parameters from file}
\usage{
load_parameters_from_file(filename)
}
\arguments{
\item{filename}{name of the parameter file}
}
\value{
a data frame with all parameters
}
\description{
Load parameters from file
}
\author{
Richel Bilderbeek
}

