% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_rooted_branch_score.R
\name{normalized_rooted_branch_score}
\alias{normalized_rooted_branch_score}
\title{calculates the normalized rooted branch score}
\usage{
normalized_rooted_branch_score(filenames)
}
\arguments{
\item{x, }{y}
}
\value{
the normalized rooted branch score
}
\description{
calculates the normalized rooted branch score
}
\author{
Femke Thon
}

