% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrbs.R
\name{nrbs}
\alias{nrbs}
\title{Calculates the Normalized Rooted Branch Score (NRBS)}
\usage{
nrbs(phylogeny1, phylogeny2)
}
\arguments{
\item{phylogeny1}{the first phylogeny of type 'phylo'}

\item{phylogeny2}{the other phylogeny of type 'phylo'}
}
\value{
the normalized rooted branch score
}
\description{
Calculates the Normalized Rooted Branch Score (NRBS)
}
\author{
Richel Bilderbeek
}

