% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_indices.R
\name{p2i}
\alias{p2i}
\title{Convert a posterior position to an index}
\usage{
p2i(sti, ai, pi, nstpist, napst, nppa)
}
\arguments{
\item{sti}{the species tree index, a value from 1 to and including 2}

\item{ai}{the alignment index,
the ai-th alignment of that species tree,
a value from 1 to and including
the number of alignments per species tree}

\item{pi}{the posterior index,
the pi-th psoterior of that alignment,
a value from 1 to and including
the number of posteriors per alignment}

\item{nstpist}{the number of species trees per incipient species tree}

\item{napst}{the number of alignments per species tree}

\item{nppa}{the number of posteriors per alignment}
}
\value{
the plain index
}
\description{
Convert a posterior position to an index
}
\author{
Richel Bilderbeek
}

