% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alignments.R
\name{plot_alignments}
\alias{plot_alignments}
\title{Plot alignments}
\usage{
plot_alignments(filename)
}
\arguments{
\item{filename}{name of the parameter file}
}
\description{
Plot alignments
}
\author{
Richel Bilderbeek
}

