% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_tree_versus_species_trees.R
\name{plot_gene_tree_versus_species_trees}
\alias{plot_gene_tree_versus_species_trees}
\title{Plots the incipient species tree and species tree}
\usage{
plot_gene_tree_versus_species_trees()
}
\description{
Plots the incipient species tree and species tree
}

