% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_nltts.R
\name{plot_posterior_nltts}
\alias{plot_posterior_nltts}
\title{Plot multiple BEAST2 posteriors}
\usage{
plot_posterior_nltts(filename)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}
}
\value{
Nothing, produces a plot
}
\description{
Plot multiple BEAST2 posteriors
}

