% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_sample_nltts.R
\name{plot_posterior_sample_nltts}
\alias{plot_posterior_sample_nltts}
\title{Plot sergpwe9grgerg}
\usage{
plot_posterior_sample_nltts(filename)
}
\arguments{
\item{filename}{a filename}
}
\description{
Plot sergpwe9grgerg
}

