% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_samples.R
\name{plot_posterior_samples}
\alias{plot_posterior_samples}
\title{Plot}
\usage{
plot_posterior_samples(filename)
}
\arguments{
\item{filename}{a filename}
}
\value{
Nothing, but it does generate some plots
}
\description{
Plot
}

