% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_tree.R
\name{plot_species_tree}
\alias{plot_species_tree}
\title{Plot}
\usage{
plot_species_tree(filename)
}
\arguments{
\item{filename}{a file name}
}
\value{
Nothing, but it does generate plots
}
\description{
Plot
}
\author{
Richel Bilderbeek
}

