% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_tree_nltt.R
\name{plot_species_tree_nltt}
\alias{plot_species_tree_nltt}
\title{Plot species_tree_with_outgroup_nltt}
\usage{
plot_species_tree_nltt(filename, dt = 0.001)
}
\arguments{
\item{filename}{parameter filename}

\item{dt}{the resolution of the nLTT plot,
must be in range <0,1>, default is 0.001}
}
\description{
Plot species_tree_with_outgroup_nltt
}
\author{
Richel Bilderbeek
}

