% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_tree_with_outgroup.R
\name{plot_species_tree_with_outgroup}
\alias{plot_species_tree_with_outgroup}
\title{Plot}
\usage{
plot_species_tree_with_outgroup(filename)
}
\arguments{
\item{filename}{a file name}
}
\value{
Nothing, but it does generate plots
}
\description{
Plot
}
\author{
Richel Bilderbeek
}

