% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_tree_with_outgroup_nltt.R
\name{plot_species_tree_with_outgroup_nltt}
\alias{plot_species_tree_with_outgroup_nltt}
\title{Plot species_tree_with_outgroup_nltt}
\usage{
plot_species_tree_with_outgroup_nltt(filename)
}
\arguments{
\item{filename}{parameter filename}
}
\description{
Plot species_tree_with_outgroup_nltt
}

