% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors_same_distr.R
\name{posteriors_same_distr}
\alias{posteriors_same_distr}
\title{Tests if the nLTT statistics are of the same distribution}
\usage{
posteriors_same_distr(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame,
}
\description{
Tests if the nLTT statistics are of the same distribution
}

