% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_esses.R
\name{read_collected_esses}
\alias{read_collected_esses}
\title{Read all the collected ESSes of all simulations' posteriors}
\usage{
read_collected_esses()
}
\value{
a dataframe
}
\description{
Read all the collected ESSes of all simulations' posteriors
}
\examples{
  df <- read_collected_esses()
  expected_names <- c("filename", "sti", "ai", "pi", "min_ess")
  testit::assert(names(df) == expected_names)
  testit::assert(is.factor(df$filename))
  testit::assert(is.factor(df$sti))
  testit::assert(is.factor(df$ai))
  testit::assert(is.factor(df$pi))
  testit::assert(is.factor(df$si))
}
\author{
Richel Bilderbeek
}

