% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_nltt_stats.R
\name{read_collected_nltt_stats}
\alias{read_collected_nltt_stats}
\title{Read all the collected nLTT statistics of all simulations}
\usage{
read_collected_nltt_stats()
}
\value{
a dataframe
}
\description{
Read all the collected nLTT statistics of all simulations
}
\examples{
  df <- read_collected_nltt_stats()
  expected_names <- c("filename", "sti", "ai", "pi", "si", "nltt_stat")
  testit::assert(names(df) == expected_names)
  testit::assert(is.factor(df$filename))
  testit::assert(is.factor(df$sti))
  testit::assert(is.factor(df$ai))
  testit::assert(is.factor(df$pi))
  testit::assert(is.factor(df$si))
}
\author{
Richel Bilderbeek
}

