% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_parameters.R
\name{read_collected_parameters}
\alias{read_collected_parameters}
\title{Reads the file with the collected parameters}
\usage{
read_collected_parameters()
}
\value{
a data frame with all parameters
}
\description{
Reads the file with the collected parameters
}
\author{
Richel Bilderbeek
}

