% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collected_time.R
\name{read_collected_times}
\alias{read_collected_times}
\title{Reads the file with the collected times}
\usage{
read_collected_times()
}
\value{
a list containing two data frames,
  'per_file' and 'full_process', with the time measurements
  per file and for the full process.
}
\description{
Reads the file with the collected times
}
\author{
Richel Bilderbeek
}

