% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read a parameter (.RDa) file}
\usage{
read_file(filename)
}
\arguments{
\item{filename}{name of the parameter (.RDa) file}
}
\value{
the file as R data
}
\description{
Read a parameter (.RDa) file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  print(names(file))
}
\author{
Richel Bilderbeek
}

