% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_species_trees.R
\name{sample_species_trees}
\alias{sample_species_trees}
\title{Samples species trees from PBD::pbd_sim its output
It does not use pbd_sim()$stree, but generates these like PBD does.}
\usage{
sample_species_trees(n, pbd_sim_output)
}
\arguments{
\item{n}{The number of species trees to sample}

\item{pbd_sim_output}{The output generated by PBD::pbd_sim}
}
\value{
A list of length n, containing the n sampled species trees
}
\description{
Samples species trees from PBD::pbd_sim its output
It does not use pbd_sim()$stree, but generates these like PBD does.
}
\author{
Richel Bilderbeek, inspired by Rampal Etienne's PBD::pbd_sim function
}

