% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_parameters_to_file.R
\name{save_parameters_to_file}
\alias{save_parameters_to_file}
\title{Creates a valid parameter file}
\usage{
save_parameters_to_file(rng_seed, sirg, siri, scr, erg, eri, age,
  n_species_trees_samples, mutation_rate, n_alignments, sequence_length,
  mcmc_chainlength, n_beast_runs, filename)
}
\arguments{
\item{rng_seed}{the random number generator seed}

\item{sirg}{the species initiation rate of the good species}

\item{siri}{the speciation-initiation rate of the incipient species}

\item{scr}{the rate at which incipient species become good}

\item{erg}{the rate at which good species get extinct}

\item{eri}{the rate at which incipient species get extinct}

\item{age}{crown age of the phylogeny}

\item{n_species_trees_samples}{the number of species trees that will be sampled from an incipient species tree}

\item{mutation_rate}{the probability per nucleotide to mutate at a DNA replication}

\item{n_alignments}{the number of alignments simulated per species tree}

\item{sequence_length}{the simulated DNA sequence length in nucleotides}

\item{mcmc_chainlength}{the length of the MCMC chain that BEAST2 will run}

\item{n_beast_runs}{the number of BEAST2 runs per DNA alignments}

\item{filename}{the name of the parameter file that will be created by this function}
}
\value{
Nothing, it will create a file with filename `filename`
}
\description{
Creates a valid parameter file
}
\author{
Richel Bilderbeek
}

