% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_alignment.R
\name{set_alignment}
\alias{set_alignment}
\title{Set an alignment from an index from a file}
\usage{
set_alignment(file, sti, ai, alignment)
}
\arguments{
\item{file}{A loaded parameter file}

\item{sti}{the species tree index, a value from 1 to and including 2}

\item{ai}{the alignment index,
the ai-th alignment of that species tree,
a value from 1 to and including
the number of alignments per species tree}

\item{alignment}{an alignment, may also be NA}
}
\value{
the modified file
}
\description{
Set an alignment from an index from a file
}
\author{
Richel Bilderbeek
}

