% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_alignment.R
\name{set_alignment_by_index}
\alias{set_alignment_by_index}
\title{Set an alignment of a file}
\usage{
set_alignment_by_index(file, i, alignment)
}
\arguments{
\item{file}{A loaded parameter file}

\item{i}{the index of the alignment}

\item{alignment}{an alignment, may also be NA}
}
\value{
the modified file
}
\description{
Set an alignment of a file
}
\author{
Richel Bilderbeek
}

