% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_posterior.R
\name{set_posterior}
\alias{set_posterior}
\title{Set an posterior to a file}
\usage{
set_posterior(file, sti, ai, pi, posterior)
}
\arguments{
\item{file}{A loaded parameter file}

\item{sti}{the species tree index, a value from 1 to and including 2}

\item{ai}{the alignment index,
the ai-th alignment of that species tree,
a value from 1 to and including
the number of alignments per species tree}

\item{pi}{the posterior index,
the pi-th posterior of that alignment,
a value from 1 to and including
the number of posteriors per alignment}

\item{posterior}{a posterior, may also be NA}
}
\value{
the modified file
}
\description{
Set an posterior to a file
}
\author{
Richel Bilderbeek
}

