% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{set_posterior_by_index}
\alias{set_posterior_by_index}
\title{Set a BEAST2 posterior of a file}
\usage{
set_posterior_by_index(file, posterior_index, posterior)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
nothing
}
\description{
Set a BEAST2 posterior of a file
}
\author{
Richel Bilderbeek
}

