% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_species_tree.R
\name{set_species_tree_by_index}
\alias{set_species_tree_by_index}
\title{Set an species_tree of a file}
\usage{
set_species_tree_by_index(file, sti, species_tree)
}
\arguments{
\item{file}{A loaded parameter file}

\item{sti}{the index of the species_tree}

\item{species_tree}{an species_tree, may also be NA}
}
\value{
the modified file
}
\description{
Set an species_tree of a file
}
\author{
Richel Bilderbeek
}

