% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_parameter_files.R
\name{show_parameter_files}
\alias{show_parameter_files}
\title{Creates a nice knitr::table to show one or more parameter files}
\usage{
show_parameter_files(filenames)
}
\arguments{
\item{filenames}{names of the parameter files}
}
\description{
Creates a nice knitr::table to show one or more parameter files
}
\author{
Richel Bilderbeek
}

