% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_indices.R
\name{a2i}
\alias{a2i}
\title{Convert an alignment position to an index}
\usage{
a2i(sti, ai, nstpist, napst)
}
\arguments{
\item{sti}{the species tree index, a value from 1 to and including 2}

\item{ai}{the alignment index,
the ai-th alignment of that species tree,
a value from 1 to and including
the number of alignments per species tree}

\item{nstpist}{the number of species trees per incipient species tree}

\item{napst}{the number of alignments per species tree}
}
\value{
the plain index
}
\description{
Convert an alignment position to an index
}
\examples{
  # only index must be one
  testit::assert(1 == a2i(sti = 1, ai = 1, nstpist = 1, napst = 1))

  # With only two species tree indices must be one and two
  testit::assert(1 == a2i(sti = 1, ai = 1, nstpist = 1, napst = 1))
  testit::assert(2 == a2i(sti = 2, ai = 1, nstpist = 2, napst = 1))

  # With only two alignment indices must be one and two
  testit::assert(1 == a2i(sti = 1, ai = 1, nstpist = 1, napst = 1))
  testit::assert(2 == a2i(sti = 1, ai = 2, nstpist = 1, napst = 2))

  # With only three species tree indices must be one to three
  testit::assert(1 == a2i(sti = 1, ai = 1, nstpist = 1, napst = 1))
  testit::assert(2 == a2i(sti = 2, ai = 1, nstpist = 2, napst = 1))
  testit::assert(3 == a2i(sti = 3, ai = 1, nstpist = 3, napst = 1))
}
\author{
Richel Bilderbeek
}

