% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pbd_output.R
\name{add_pbd_output}
\alias{add_pbd_output}
\title{Adds a pbd_sim result to a file}
\usage{
add_pbd_output(filename, verbose = FALSE)
}
\arguments{
\item{filename}{Parameter filename}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
Nothing, modifies the parameter file
}
\description{
Adds a pbd_sim result to a file
}
\examples{
  # Create a parameter file
  filename <- "add_pbd_output_example.RDa"
  save_parameters_to_file(
    rng_seed = 42,
    sirg = 0.5,
    siri = 0.5,
    scr = 0.5,
    erg = 0.5,
    eri = 0.5,
    age = 5,
    mutation_rate = 0.1,
    n_alignments = 1,
    sequence_length = 10,
    nspp = 10,
    n_beast_runs = 1,
    filename = filename
  )

  # File does not have an incipient species tree yet
  testit::assert(!is_pbd_sim_output(read_file(filename)$pbd_output))

  # Add an incipient species tree
  add_pbd_output(filename)

  # File does have an incipient species tree now
  testit::assert(is_pbd_sim_output(read_file(filename)$pbd_output))

  # Cleanup
  file.remove(filename)
}
\author{
Richel Bilderbeek
}

