% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posteriors.R
\name{add_posteriors}
\alias{add_posteriors}
\title{Add BEAST2 posteriors to a file}
\usage{
add_posteriors(filename, skip_if_output_present = FALSE, verbose = TRUE)
}
\arguments{
\item{filename}{Parameter filename}

\item{skip_if_output_present}{skip if output files are present, else remove these and start a new BEAST2 run}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
the number of posteriors added. Also modifies the parameter file
}
\description{
Add BEAST2 posteriors to a file
}
\author{
Richel Bilderbeek
}

