% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_species_trees.R
\name{add_species_trees}
\alias{add_species_trees}
\title{Add a species tree with/without outgroup to a file}
\usage{
add_species_trees(filename)
}
\arguments{
\item{filename}{Parameter filename}
}
\value{
Nothing, modifies the parameter file
}
\description{
Add a species tree with/without outgroup to a file
}
\author{
Richel Bilderbeek
}

