% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_to_beast_posterior.R
\name{alignment_to_beast_posterior}
\alias{alignment_to_beast_posterior}
\title{Convert an alignment to a BEAST2 XML input file}
\usage{
alignment_to_beast_posterior(alignment_dnabin, nspp, base_filename,
  rng_seed = 42, beast_jar_path = find_beast_jar_path(),
  skip_if_output_present = FALSE, verbose = FALSE)
}
\arguments{
\item{alignment_dnabin}{The alignmnet}

\item{nspp}{The number of states in the MCMC chain BEAST2 will generate,
typically one state per one thousand moves}

\item{base_filename}{The base of the filename (the part without the extension)}

\item{rng_seed}{The random number generator seed used by BEAST2}

\item{beast_jar_path}{Where the jar 'beast.jar' can be found}

\item{skip_if_output_present}{skip if output files are present, else remove these and start a new BEAST2 run}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
the phylogenies of the BEAST2 posterior
}
\description{
Convert an alignment to a BEAST2 XML input file
}
\examples{

  # Prepare the many filenames
  base_filename <- "alignment_to_beast_posterior_example"
  beast_log_filename <- paste0(base_filename, ".log")
  beast_trees_filename <- paste0(base_filename, ".trees")
  beast_state_filename <- paste0(base_filename, ".xml.state")

  # Simulate a random phylogeny its alignment
  alignment <- convert_phylogeny_to_alignment(
    phylogeny = ape::rcoal(5),
    sequence_length = 10,
    mutation_rate = 1
  )

  # See if the BEAST2 .jar file is present
  beast_jar_path <- find_beast_jar_path()
  testit::assert(file.exists(beast_jar_path))

  # Run BEAST2 and extract the phylogenies of its posterior
  posterior <- alignment_to_beast_posterior(
    alignment_dnabin = alignment,
    nspp = 10,
    base_filename = base_filename,
    rng_seed = 42,
    beast_jar_path = beast_jar_path,
    skip_if_output_present = FALSE,
    verbose = FALSE
  )

  # Check the posterior
  testit::assert(RBeast::is_posterior(posterior))
  testit::assert(RBeast::is_trees_posterior(posterior$trees))

}
\author{
Richel Bilderbeek
}

