% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_posteriors.R
\name{are_identical_posteriors}
\alias{are_identical_posteriors}
\title{Determines if the input is a Cer2016 posterior,}
\usage{
are_identical_posteriors(p, q)
}
\arguments{
\item{p}{the first Cer2016 posterior}

\item{q}{the second Cer2016 posterior}
}
\value{
TRUE or FALSE
}
\description{
Determines if the input is a Cer2016 posterior,
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  posterior_1 <- get_posterior(file = file, sti = 1, ai = 1, pi = 1)
  posterior_2 <- get_posterior(file = file, sti = 2, ai = 1, pi = 1)
  testit::assert( are_identical_posteriors(posterior_1, posterior_1))
  testit::assert(!are_identical_posteriors(posterior_1, posterior_2))
  testit::assert(!are_identical_posteriors(posterior_2, posterior_1))
  testit::assert( are_identical_posteriors(posterior_2, posterior_2))
}
\author{
Richel Bilderbeek
}

