% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_sorter.R
\name{branch_sorter}
\alias{branch_sorter}
\title{Sorts branches for the NRBS analysis}
\usage{
branch_sorter(phylogeny)
}
\arguments{
\item{phylogeny}{A phylogeny of class 'phylo'}
}
\value{
something
}
\description{
Sorts branches for the NRBS analysis
}
\author{
Femke Thon
}

