% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nltt_stats.R
\name{calc_nltt_stats}
\alias{calc_nltt_stats}
\title{Calculate the nLTT statistic between a focal phylogeny
and multiple others}
\usage{
calc_nltt_stats(phylogeny, others)
}
\arguments{
\item{phylogeny}{the focal phylogeny}

\item{others}{the other phylogenies}
}
\value{
A data frame
}
\description{
Calculate the nLTT statistic between a focal phylogeny
and multiple others
}
\examples{
  nltt_stats <- calc_nltt_stats(
    phylogeny = ape::rcoal(10),
    others = c(ape::rcoal(10), ape::rcoal(10))
  )
  testit::assert(names(nltt_stats) == c("id", "nltt_stat"))
  testit::assert(is.factor(nltt_stats$id))
}
\author{
Richel Bilderbeek
}

