% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_gammas.R
\name{collect_files_gammas}
\alias{collect_files_gammas}
\title{Collects the gamma statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_gammas(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A list with two dataframes of gamma statistics
}
\description{
Collects the gamma statistics of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
\examples{
 filenames <- c(
   find_path("toy_example_3.RDa"),
   find_path("toy_example_4.RDa")
 )
 df <- collect_files_gammas(filenames)
 testit::assert(
   names(df) ==
   c("species_tree_gamma_stats", "posterior_gamma_stats")
 )
 testit::assert(
   names(df$species_tree_gamma_stats) ==
   c("filename", "species_tree", "gamma_stat")
 )
 testit::assert(nrow(df$species_tree_gamma_stats) == 4)
 testit::assert(nrow(df$posterior_gamma_stats) == 160)
}

