% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_files_n_taxa.R
\name{collect_files_n_taxa}
\alias{collect_files_n_taxa}
\title{Collects the number of taxa of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form}
\usage{
collect_files_n_taxa(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter file}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
A dataframe with all number of taxa of all files
}
\description{
Collects the number of taxa of all phylogenies belonging to a
multiple parameter file in the melted/uncast/long form
}
\examples{
  filenames <- c(
   find_path("toy_example_1.RDa"),
   find_path("toy_example_2.RDa")
 )
 df <- collect_files_n_taxa(filenames, verbose = FALSE)
 testit::assert(names(df) == c("filename", "n_taxa"))
 testit::assert(nrow(df) == length(filenames))
}

