% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_parameters.R
\name{collect_parameters}
\alias{collect_parameters}
\title{Creates tidy data of all parameter files}
\usage{
collect_parameters(filenames, verbose = FALSE)
}
\arguments{
\item{filenames}{names of the parameter files}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\description{
Creates tidy data of all parameter files
}
\examples{
 filenames <- c(
   find_path("article_0_0_0_0_0.RDa"),
   find_path("article_0_1_4_0_2.RDa"),
   find_path("toy_example_1.RDa")
 )
 df <- collect_parameters(filenames)
 testit::assert(nrow(df) == 3)
}
\author{
Richel Bilderbeek
}

