% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_posterior_filenames.R
\name{collect_posterior_filenames}
\alias{collect_posterior_filenames}
\title{Collect posterior filenames}
\usage{
collect_posterior_filenames(parameter_filename)
}
\arguments{
\item{filename}{Parameter filename}
}
\value{
generates species tree files from the posterior
}
\description{
Collect posterior filenames
}
\author{
Jolien Gay
}

