% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_posterior_gammas.R
\name{collect_posterior_gammas}
\alias{collect_posterior_gammas}
\title{Collect the gamma statistics of the BEAST2 posteriors}
\usage{
collect_posterior_gammas(filename)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}
}
\value{
a data frame
}
\description{
Collect the gamma statistics of the BEAST2 posteriors
}
\examples{
  filename <- find_path("toy_example_3.RDa")
  df <- collect_posterior_gammas(filename)
  testit::assert(names(df) ==
    c("species_tree", "alignment", "beast_run", "gamma")
  )
  testit::assert(nrow(df) == 80)
}
\author{
Richel Bilderbeek
}

