% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_species_tree_gammas.R
\name{collect_species_tree_gammas}
\alias{collect_species_tree_gammas}
\title{Collect the gamma statistics of the species trees with outgroup}
\usage{
collect_species_tree_gammas(filename)
}
\arguments{
\item{filename}{name of the file containing the parameters and results}
}
\value{
a data frame
}
\description{
Collect the gamma statistics of the species trees with outgroup
}
\examples{
 filename <- find_path("toy_example_1.RDa")
 df <- collect_species_tree_gammas(filename)
 testit::assert(names(df) == c("species_tree", "gamma_stat"))
 testit::assert(nrow(df) == 1)
 testit::assert(abs(df$gamma_stat - -0.8205514) < 0.0001)
}
\author{
Richel Bilderbeek
}

