% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_parameter_files.R
\name{create_test_parameter_files}
\alias{create_test_parameter_files}
\title{Creates the parameter files for the toy examples}
\usage{
create_test_parameter_files(filenames = paste0("toy_example_", seq(1, 4),
  ".RDa"))
}
\arguments{
\item{filenames}{the name of the four files created}
}
\value{
nothing
}
\description{
Creates the parameter files for the toy examples
}
\author{
Richel Bilderbeek
}

