% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_beast2_paths.R
\name{find_beast_bin_path}
\alias{find_beast_bin_path}
\title{Find the path of the BEAST2 binary file}
\usage{
find_beast_bin_path()
}
\value{
the path of the BEAST2 binary file
}
\description{
Find the path of the BEAST2 binary file
}
\examples{
  my_file <- find_beast2_bin_path()
  does_exist <- file.exists(my_file)
  testit::assert(does_exist)
}

