% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_paths.R
\name{find_beast_jar_path}
\alias{find_beast_jar_path}
\title{Find the path of the BEAST2 jar file}
\usage{
find_beast_jar_path()
}
\value{
the path of the BEAST2 jar file
}
\description{
Find the path of the BEAST2 jar file
}

