% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylogeny_crown_age.R
\name{get_phylogeny_crown_age}
\alias{get_phylogeny_crown_age}
\title{Obtain the crown age of a phylony}
\usage{
get_phylogeny_crown_age(phylogeny)
}
\arguments{
\item{phylogeny}{The phylogeny to obtain the crown age of}
}
\value{
the age of the phylogeny
}
\description{
Obtain the crown age of a phylony
}
\examples{
  age <- 15
  set.seed(42)
  phylogeny <- PBD::pbd_sim(
    c(0.2, 1, 0.2, 0.0, 0.0), age
  )$tree
  n_taxa <- length(phylogeny$tip.label)
  testit::assert(n_taxa > 0)
  crown_age <- get_phylogeny_crown_age(phylogeny)
  testit::assert(all.equal(age, crown_age, tolerance = 0.001))
}
\author{
Richel Bilderbeek
}

