% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{get_posterior}
\alias{get_posterior}
\title{Extract a BEAST2 posterior from a file}
\usage{
get_posterior(file, sti, ai, pi)
}
\arguments{
\item{file}{A loaded parameter file}

\item{sti}{the species tree index, a value from 1 to and including 2}

\item{ai}{the alignment index,
the ai-th alignment of that species tree,
a value from 1 to and including
the number of alignments per species tree}

\item{pi}{the posterior index,
the pi-th posterior of that alignment,
a value from 1 to and including
the number of posteriors per alignment}
}
\value{
the posterior
}
\description{
Extract a BEAST2 posterior from a file
}
\examples{
  # Read a file with one or more posteriors
  file <- read_file(find_path("toy_example_1.RDa"))
  # Pick the indices of the posterior to extract
  sti  <- 1 # Species tree index
  ai   <- 1 # Alignment index
  pi   <- 1 # Posterior index
  # Extract the posterior
  posterior <- get_posterior(file = file, sti = sti, ai = 1, pi = 1)
  # Check that it is indeed a posterior of non-zero length
  testit::assert(all(names(posterior) == c("trees", "estimates")))
  testit::assert(RBeast::is_posterior(posterior))
  testit::assert(length(posterior) > 0)
}
\author{
Richel Bilderbeek
}

