% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{get_posterior_by_index}
\alias{get_posterior_by_index}
\title{Extract a BEAST2 posterior phyogenies
  from a file}
\usage{
get_posterior_by_index(file, i)
}
\arguments{
\item{file}{A loaded parameter file}

\item{i}{the posterior index,
the index of the posteriors in the list of posteriors}
}
\value{
the posterior
}
\description{
Extract a BEAST2 posterior phyogenies
  from a file
}
\author{
Richel Bilderbeek
}

