% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_file.R
\name{is_valid_file}
\alias{is_valid_file}
\title{Checks if a file is a valid parameter file}
\usage{
is_valid_file(filename, verbose = FALSE)
}
\arguments{
\item{filename}{the name of the file to be checked}

\item{verbose}{give verbose output, should be TRUE or FALSE}
}
\value{
TRUE or FALSE
}
\description{
Checks if a file is a valid parameter file
}
\examples{
  testit::assert(is_valid_file(find_path("toy_example_1.RDa")))
}
\author{
Richel Bilderbeek
}

