% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrbs.R
\name{nrbs}
\alias{nrbs}
\title{Calculates the Normalized Rooted Branch Score (NRBS)}
\usage{
nrbs(phylogeny1, phylogeny2)
}
\arguments{
\item{phylogeny1}{the first phylogeny of type 'phylo'}

\item{phylogeny2}{the other phylogeny of type 'phylo'}
}
\value{
the normalized rooted branch score
}
\description{
Calculates the Normalized Rooted Branch Score (NRBS)
}
\details{
The NRBS 'is the rooted equivalent of the branch score metric suggested
  by Kuhner and Felsenstein' (quoted from Heled & Drummond, 2010).
  phangorn::KF.dist returns by default the unrooted value,
  but can be changed to return the rooted value
}
\author{
Richel Bilderbeek
}
\references{
Heled, Joseph, and Alexei J. Drummond. "Bayesian inference of species trees from multilocus data." Molecular biology and evolution 27.3 (2010): 570-580.
  Kuhner, Mary K., and Joseph Felsenstein. "A simulation comparison of phylogeny algorithms under equal and unequal evolutionary rates." Molecular Biology and Evolution 11.3 (1994): 459-468.
}

