% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extant_incipient_tree.R
\name{plot_extant_incipient_tree}
\alias{plot_extant_incipient_tree}
\title{Plot the extant inicipient species tree}
\usage{
plot_extant_incipient_tree(filename)
}
\arguments{
\item{filename}{name of the parameter file}
}
\value{
Nothing. It does produce a plot
}
\description{
Plot the extant inicipient species tree
}

