% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_sample_nltts.R
\name{plot_posterior_sample_nltts}
\alias{plot_posterior_sample_nltts}
\title{Plot the nLTTs of the posterior}
\usage{
plot_posterior_sample_nltts(filename, dt = 0.001)
}
\arguments{
\item{filename}{a filename}

\item{dt}{the resolution of the nLTT plot,
must be in range <0,1>, default is 0.001}
}
\value{
Nothing, but it does generate some plots
}
\description{
Plot the nLTTs of the posterior
}
\author{
Richel Bilderbeek
}

