% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_sample_nltts.R
\name{plot_posterior_sample_nltts}
\alias{plot_posterior_sample_nltts}
\title{Plot the nLTTs of the posterior}
\usage{
plot_posterior_sample_nltts(filename)
}
\arguments{
\item{filename}{a filename}
}
\value{
Nothing, but it does generate some plots
}
\description{
Plot the nLTTs of the posterior
}
\author{
Richel Bilderbeek
}

