% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_parameters_to_file.R
\name{save_parameters_to_file}
\alias{save_parameters_to_file}
\title{Creates a valid parameter file}
\usage{
save_parameters_to_file(rng_seed, sirg, siri, scr, erg, eri, age, mutation_rate,
  n_alignments, sequence_length, nspp, n_beast_runs, filename)
}
\arguments{
\item{rng_seed}{the random number generator seed}

\item{sirg}{the species initiation rate of the good species}

\item{siri}{the speciation-initiation rate of the incipient species}

\item{scr}{the rate at which incipient species become good}

\item{erg}{the rate at which good species get extinct}

\item{eri}{the rate at which incipient species get extinct}

\item{age}{crown age of the phylogeny}

\item{mutation_rate}{the probability per nucleotide to mutate at a DNA replication}

\item{n_alignments}{the number of alignments simulated per species tree}

\item{sequence_length}{the simulated DNA sequence length in nucleotides}

\item{nspp}{the number of states in the MCMC chain that BEAST2 will run}

\item{n_beast_runs}{the number of BEAST2 runs per DNA alignments}

\item{filename}{the name of the parameter file that will be created by this function}
}
\value{
Nothing, it will create a file with filename `filename`
}
\description{
Creates a valid parameter file
}
\examples{
filename <- "save_parameters_to_file_example.RDa"
save_parameters_to_file(
  rng_seed = 42,
  sirg = 0.5,
  siri = 0.5,
  scr = 0.5,
  erg = 0.5,
  eri = 0.5,
  age = 5,
  mutation_rate = 0.1,
  n_alignments = 1,
  sequence_length = 10,
  nspp = 10,
  n_beast_runs = 1,
  filename = filename
)
testit::assert(is_valid_file(filename) == TRUE)
}
\author{
Richel Bilderbeek
}

