% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_posterior.R
\name{set_posterior_by_index}
\alias{set_posterior_by_index}
\title{Set a BEAST2 posterior of a file}
\usage{
set_posterior_by_index(file, i, posterior)
}
\arguments{
\item{file}{A loaded parameter file}

\item{i}{the index of the posterior}

\item{posterior}{a BEAST2 posterior, may also be NA}
}
\value{
the modified file
}
\description{
Set a BEAST2 posterior of a file
}
\author{
Richel Bilderbeek
}

